Flatter is better
